chore(core/rust): make TextBox common component
