+++ core/embed/rust/src/ui/component/text/mod.rs
@@ -1,5 +1,5 @@
+pub mod common;
 pub mod formatted;
-mod iter;
 pub mod layout;
 pub mod paragraphs;
